requirements.txt
aiohttp==3.9.5
altgraph==0.17
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
dict==2020.12.3
discord.py==2.4.0
discord.py[voice]==2.4.0
discord-py-slash-command==3.0.3
future==0.18.2
idna==2.10
multidict==5.1.0
pefile==2019.4.18
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyNaCl==1.4.0
python-dateutil==1.5
pytz==2024.1
pywin32-ctypes==0.2.0
requests==2.32.3
six==1.16.0
spotipy==2.24.0
typing-extensions==3.7.4.3
urllib3==2.2.2
values==2020.12.3
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2026.1.31